backend/requirements.txt
sentence-transformers>=2.2.0
numpy>=1.24.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic[email]>=2.0.0
pytest>=7.4.0
httpx>=0.25.0
pytest-asyncio>=0.21.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
authlib>=1.3.0